requirements.txt
Flask==3.0.0
web3==6.11.3
python-dotenv==1.0.0
eth-utils==2.3.1